+++ src/entry.rs
@@ -17,16 +17,17 @@
 use std::marker::PhantomData;
 
 use super::{
-    base_index::BaseIndex, indexes_metadata::IndexType, Fork, Snapshot, StorageKey, StorageValue,
+    base_index::BaseIndex, indexes_metadata::IndexType, BinaryKey, BinaryValue, Fork, Snapshot,
+    UniqueHash,
 };
 use exonum_crypto::Hash;
 
 /// An index that may only contain one element.
 ///
 /// You can add an element to this index and check whether it exists. A value
-/// should implement [`StorageValue`] trait.
+/// should implement [`BinaryValue`] trait.
 ///
-/// [`StorageValue`]: trait.StorageValue.html
+/// [`BinaryValue`]: trait.BinaryValue.html
 #[derive(Debug)]
 pub struct Entry<T, V> {
     base: BaseIndex<T>,
@@ -36,7 +37,7 @@ pub struct Entry<T, V> {
 impl<T, V> Entry<T, V>
 where
     T: AsRef<dyn Snapshot>,
-    V: StorageValue,
+    V: BinaryValue + UniqueHash,
 {
     /// Creates a new index representation based on the name and storage view.
     ///
@@ -87,7 +88,7 @@ where
     /// ```
     pub fn new_in_family<S, I>(family_name: S, index_id: &I, view: T) -> Self
     where
-        I: StorageKey,
+        I: BinaryKey,
         I: ?Sized,
         S: AsRef<str>,
     {
@@ -165,7 +166,7 @@ where
 
 impl<'a, V> Entry<&'a mut Fork, V>
 where
-    V: StorageValue,
+    V: BinaryValue + UniqueHash,
 {
     /// Changes a value of the entry.
     ///