+++ src/values.rs
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! A definition of `StorageValue` trait and implementations for common types.
+//! A definition of `BinaryValue` trait and implementations for common types.
 
-use std::{borrow::Cow, mem};
+use std::{borrow::Cow, io::Read};
 
-use byteorder::{ByteOrder, LittleEndian};
+use byteorder::{ByteOrder, LittleEndian, ReadBytesExt};
 use chrono::{DateTime, NaiveDateTime, Utc};
+use failure::{self, format_err};
 use rust_decimal::Decimal;
 use uuid::Uuid;
 
@@ -26,19 +27,18 @@ use exonum_crypto::{Hash, PublicKey};
 
 /// A type that can be (de)serialized as a value in the blockchain storage.
 ///
-/// `StorageValue` is automatically implemented by the [`encoding_struct!`] and [`transactions!`]
-/// macros. In case you need to implement it manually, use little-endian encoding
+/// If you need to implement `BinaryValue` for your types, use little-endian encoding
 /// for integer types for compatibility with modern architectures.
 ///
 /// # Examples
 ///
-/// Implementing `StorageValue` for the type:
+/// Implementing `BinaryValue` for the type:
 ///
 /// ```
-/// use std::borrow::Cow;
-/// use exonum_merkledb::StorageValue;
-/// use exonum_crypto::{self, CryptoHash, Hash};
-/// use byteorder::{LittleEndian, ByteOrder};
+/// use std::{borrow::Cow, io::{Read, Write}};
+/// use byteorder::{LittleEndian, ReadBytesExt, ByteOrder};
+/// use failure;
+/// use exonum_merkledb::BinaryValue;
 ///
 /// #[derive(Clone)]
 /// struct Data {
@@ -46,211 +46,174 @@ use exonum_crypto::{Hash, PublicKey};
 ///     b: u32,
 /// }
 ///
-/// impl CryptoHash for Data {
-///     fn hash(&self) -> Hash {
-///         let mut buffer = [0; 6];
-///         LittleEndian::write_i16(&mut buffer[0..2], self.a);
-///         LittleEndian::write_u32(&mut buffer[2..6], self.b);
-///         exonum_crypto::hash(&buffer)
-///     }
-/// }
-///
-/// impl StorageValue for Data {
-///     fn into_bytes(self) -> Vec<u8> {
-///         let mut buffer = vec![0; 6];
-///         LittleEndian::write_i16(&mut buffer[0..2], self.a);
-///         LittleEndian::write_u32(&mut buffer[2..6], self.b);
-///         buffer
+/// impl BinaryValue for Data {
+///     fn to_bytes(&self) -> Vec<u8> {
+///         let mut buf = vec![0_u8; 6];
+///         LittleEndian::write_i16(&mut buf[0..2], self.a);
+///         LittleEndian::write_u32(&mut buf[2..6], self.b);
+///         buf
 ///     }
 ///
-///     fn from_bytes(value: Cow<[u8]>) -> Self {
-///         let a = LittleEndian::read_i16(&value[0..2]);
-///         let b = LittleEndian::read_u32(&value[2..6]);
-///         Data { a, b }
+///     fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+///         let mut buf = bytes.as_ref();
+///         let a = buf.read_i16::<LittleEndian>()?;
+///         let b = buf.read_u32::<LittleEndian>()?;
+///         Ok(Self { a, b })
 ///     }
 /// }
 /// # fn main() {}
 /// ```
-///
-/// [`encoding_struct!`]: ../macro.encoding_struct.html
-/// [`transactions!`]: ../macro.transactions.html
-pub trait StorageValue: UniqueHash + Sized {
-    /// Serialize a value into a vector of bytes.
-    fn into_bytes(self) -> Vec<u8>;
-
-    /// Deserialize a value from bytes.
-    fn from_bytes(value: Cow<[u8]>) -> Self;
+pub trait BinaryValue: Sized {
+    /// Serializes the given value to the vector of bytes.
+    fn to_bytes(&self) -> Vec<u8>;
+    /// Consumes and serializes the given value to the vector of bytes.
+    /// This method is faster with the wrapped values,
+    /// thus if you wouldn't use value after serialization use it.
+    fn into_bytes(self) -> Vec<u8> {
+        self.to_bytes()
+    }
+    /// Deserializes the value from the given bytes array.
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error>;
 }
+
+macro_rules! impl_binary_form_scalar {
+    ($type:tt, $read:ident) => {
+        impl BinaryValue for $type {
+            fn to_bytes(&self) -> Vec<u8> {
+                vec![*self as u8]
+            }
+
+            fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+                use byteorder::ReadBytesExt;
+                bytes.as_ref().$read().map_err(From::from)
+            }
+        }
+
+        impl UniqueHash for $type {}
+    };
+    ($type:tt, $write:ident, $read:ident, $len:expr) => {
+        impl BinaryValue for $type {
+            fn to_bytes(&self) -> Vec<u8> {
+                let mut v = vec![0; $len];
+                LittleEndian::$write(&mut v, *self);
+                v
+            }
+
+            fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+                use byteorder::ReadBytesExt;
+                bytes.as_ref().$read::<LittleEndian>().map_err(From::from)
+            }
+        }
+
+        impl UniqueHash for $type {}
+    };
+}
+
+// Unsigned scalar types
+impl_binary_form_scalar! { u8,  read_u8 }
+impl_binary_form_scalar! { u16, write_u16, read_u16, 2 }
+impl_binary_form_scalar! { u32, write_u32, read_u32, 4 }
+impl_binary_form_scalar! { u64, write_u64, read_u64, 8 }
+// Signed scalar types
+impl_binary_form_scalar! { i8,  read_i8 }
+impl_binary_form_scalar! { i16, write_i16, read_i16, 2 }
+impl_binary_form_scalar! { i32, write_i32, read_i32, 4 }
+impl_binary_form_scalar! { i64, write_i64, read_i64, 8 }
 
 /// No-op implementation.
-impl StorageValue for () {
-    fn into_bytes(self) -> Vec<u8> {
-        Vec::new()
+impl BinaryValue for () {
+    fn to_bytes(&self) -> Vec<u8> {
+        Vec::default()
     }
 
-    fn from_bytes(_value: Cow<[u8]>) -> Self {}
+    fn from_bytes(_bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        Ok(())
+    }
 }
 
-impl StorageValue for bool {
-    fn into_bytes(self) -> Vec<u8> {
-        vec![self as u8]
+impl UniqueHash for () {}
+
+impl BinaryValue for bool {
+    fn to_bytes(&self) -> Vec<u8> {
+        vec![*self as u8]
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        let value = bytes.as_ref();
         assert_eq!(value.len(), 1);
 
         match value[0] {
-            0 => false,
-            1 => true,
-            value => panic!("Invalid value for bool: {}", value),
+            0 => Ok(false),
+            1 => Ok(true),
+            value => Err(format_err!("Invalid value for bool: {}", value)),
         }
     }
 }
 
-impl StorageValue for u8 {
-    fn into_bytes(self) -> Vec<u8> {
-        vec![self]
+impl UniqueHash for bool {}
+
+impl BinaryValue for Vec<u8> {
+    fn to_bytes(&self) -> Vec<u8> {
+        self.clone()
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        assert_eq!(value.len(), 1);
-        value[0]
-    }
-}
-
-/// Uses little-endian encoding.
-impl StorageValue for u16 {
-    fn into_bytes(self) -> Vec<u8> {
-        let mut v = vec![0; 2];
-        LittleEndian::write_u16(&mut v, self);
-        v
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        LittleEndian::read_u16(value.as_ref())
-    }
-}
-
-/// Uses little-endian encoding.
-impl StorageValue for u32 {
-    fn into_bytes(self) -> Vec<u8> {
-        let mut v = vec![0; 4];
-        LittleEndian::write_u32(&mut v, self);
-        v
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        LittleEndian::read_u32(value.as_ref())
-    }
-}
-
-/// Uses little-endian encoding.
-impl StorageValue for u64 {
-    fn into_bytes(self) -> Vec<u8> {
-        let mut v = vec![0; mem::size_of::<Self>()];
-        LittleEndian::write_u64(&mut v, self);
-        v
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        LittleEndian::read_u64(value.as_ref())
-    }
-}
-
-impl StorageValue for i8 {
-    fn into_bytes(self) -> Vec<u8> {
-        vec![self as u8]
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        assert_eq!(value.len(), 1);
-        value[0] as Self
-    }
-}
-
-/// Uses little-endian encoding.
-impl StorageValue for i16 {
-    fn into_bytes(self) -> Vec<u8> {
-        let mut v = vec![0; 2];
-        LittleEndian::write_i16(&mut v, self);
-        v
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        LittleEndian::read_i16(value.as_ref())
-    }
-}
-
-/// Uses little-endian encoding.
-impl StorageValue for i32 {
-    fn into_bytes(self) -> Vec<u8> {
-        let mut v = vec![0; 4];
-        LittleEndian::write_i32(&mut v, self);
-        v
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        LittleEndian::read_i32(value.as_ref())
-    }
-}
-
-/// Uses little-endian encoding.
-impl StorageValue for i64 {
-    fn into_bytes(self) -> Vec<u8> {
-        let mut v = vec![0; 8];
-        LittleEndian::write_i64(&mut v, self);
-        v
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        LittleEndian::read_i64(value.as_ref())
-    }
-}
-
-impl StorageValue for Hash {
-    fn into_bytes(self) -> Vec<u8> {
-        self.as_ref().to_vec()
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        Self::from_slice(value.as_ref()).unwrap()
-    }
-}
-
-impl StorageValue for PublicKey {
-    fn into_bytes(self) -> Vec<u8> {
-        self.as_ref().to_vec()
-    }
-
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        Self::from_slice(value.as_ref()).unwrap()
-    }
-}
-
-impl StorageValue for Vec<u8> {
     fn into_bytes(self) -> Vec<u8> {
         self
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        value.into_owned()
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        Ok(bytes.into_owned())
     }
 }
 
-/// Uses UTF-8 string serialization.
-impl StorageValue for String {
+impl UniqueHash for Vec<u8> {}
+
+impl BinaryValue for String {
+    fn to_bytes(&self) -> Vec<u8> {
+        self.as_bytes().to_owned()
+    }
+
     fn into_bytes(self) -> Vec<u8> {
         Self::into_bytes(self)
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        Self::from_utf8(value.into_owned()).unwrap()
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        Self::from_utf8(bytes.into_owned()).map_err(From::from)
     }
 }
 
-/// Uses little-endian encoding.
-impl StorageValue for DateTime<Utc> {
-    fn into_bytes(self) -> Vec<u8> {
+impl UniqueHash for String {}
+
+impl BinaryValue for Hash {
+    fn to_bytes(&self) -> Vec<u8> {
+        self.as_ref().to_vec()
+    }
+
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        Self::from_slice(bytes.as_ref()).ok_or_else(|| {
+            format_err!("Unable to decode Hash from bytes: buffer size does not match")
+        })
+    }
+}
+
+impl BinaryValue for PublicKey {
+    fn to_bytes(&self) -> Vec<u8> {
+        self.as_ref().to_vec()
+    }
+
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        Self::from_slice(bytes.as_ref()).ok_or_else(|| {
+            format_err!("Unable to decode PublicKey from bytes: buffer size does not match")
+        })
+    }
+}
+
+impl UniqueHash for PublicKey {}
+
+// FIXME Maybe we should remove this implementations. [ECR-2775]
+
+impl BinaryValue for DateTime<Utc> {
+    fn to_bytes(&self) -> Vec<u8> {
         let secs = self.timestamp();
         let nanos = self.timestamp_subsec_nanos();
 
@@ -260,34 +223,45 @@ impl StorageValue for DateTime<Utc> {
         buffer
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        let secs = LittleEndian::read_i64(&value[0..8]);
-        let nanos = LittleEndian::read_u32(&value[8..12]);
-        Self::from_utc(NaiveDateTime::from_timestamp(secs, nanos), Utc)
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        let mut value = bytes.as_ref();
+        let secs = value.read_i64::<LittleEndian>()?;
+        let nanos = value.read_u32::<LittleEndian>()?;
+        Ok(Self::from_utc(
+            NaiveDateTime::from_timestamp(secs, nanos),
+            Utc,
+        ))
     }
 }
 
-impl StorageValue for Uuid {
-    fn into_bytes(self) -> Vec<u8> {
+impl UniqueHash for DateTime<Utc> {}
+
+impl BinaryValue for Uuid {
+    fn to_bytes(&self) -> Vec<u8> {
         self.as_bytes().to_vec()
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
-        Self::from_slice(&value).unwrap()
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        Self::from_slice(bytes.as_ref()).map_err(From::from)
     }
 }
 
-impl StorageValue for Decimal {
-    fn into_bytes(self) -> Vec<u8> {
+impl UniqueHash for Uuid {}
+
+impl BinaryValue for Decimal {
+    fn to_bytes(&self) -> Vec<u8> {
         self.serialize().to_vec()
     }
 
-    fn from_bytes(value: Cow<[u8]>) -> Self {
+    fn from_bytes(bytes: Cow<[u8]>) -> Result<Self, failure::Error> {
+        let mut value = bytes.as_ref();
         let mut buf: [u8; 16] = [0; 16];
-        buf.copy_from_slice(&value);
-        Self::deserialize(buf)
+        value.read_exact(&mut buf)?;
+        Ok(Self::deserialize(buf))
     }
 }
+
+impl UniqueHash for Decimal {}
 
 #[cfg(test)]
 mod tests {
@@ -298,88 +272,80 @@ mod tests {
 
     use super::*;
 
-    #[test]
-    fn u8_round_trip() {
-        let values = [u8::min_value(), 1, u8::max_value()];
-
-        assert_round_trip_eq(&values);
+    fn assert_round_trip_eq<T: BinaryValue + PartialEq + Debug>(values: &[T]) {
+        for value in values {
+            let bytes = value.to_bytes();
+            assert_eq!(
+                *value,
+                <T as BinaryValue>::from_bytes(bytes.into()).unwrap()
+            );
+        }
     }
 
-    #[test]
-    fn i8_round_trip() {
-        let values = [i8::min_value(), -1, 0, 1, i8::max_value()];
-
-        assert_round_trip_eq(&values);
+    macro_rules! impl_test_binary_form_scalar_unsigned {
+        ($name:ident, $type:tt) => {
+            #[test]
+            fn $name() {
+                let values = [$type::min_value(), 1, $type::max_value()];
+                assert_round_trip_eq(&values);
+            }
+        };
     }
 
-    #[test]
-    fn u16_round_trip() {
-        let values = [u16::min_value(), 1, u16::max_value()];
-
-        assert_round_trip_eq(&values);
+    macro_rules! impl_test_binary_form_scalar_signed {
+        ($name:ident, $type:tt) => {
+            #[test]
+            fn $name() {
+                let values = [$type::min_value(), -1, 0, 1, $type::max_value()];
+                assert_round_trip_eq(&values);
+            }
+        };
     }
 
-    #[test]
-    fn i16_round_trip() {
-        let values = [i16::min_value(), -1, 0, 1, i16::max_value()];
+    // Impl tests for unsigned scalar types.
+    impl_test_binary_form_scalar_unsigned! { test_binary_form_round_trip_u8,  u8 }
+    impl_test_binary_form_scalar_unsigned! { test_binary_form_round_trip_u32, u32 }
+    impl_test_binary_form_scalar_unsigned! { test_binary_form_round_trip_u16, u16 }
+    impl_test_binary_form_scalar_unsigned! { test_binary_form_round_trip_u64, u64 }
 
-        assert_round_trip_eq(&values);
-    }
+    // Impl tests for signed scalar types.
+    impl_test_binary_form_scalar_signed! { test_binary_form_round_trip_i8,  i8 }
+    impl_test_binary_form_scalar_signed! { test_binary_form_round_trip_i16, i16 }
+    impl_test_binary_form_scalar_signed! { test_binary_form_round_trip_i32, i32 }
+    impl_test_binary_form_scalar_signed! { test_binary_form_round_trip_i64, i64 }
 
-    #[test]
-    fn u32_round_trip() {
-        let values = [u32::min_value(), 1, u32::max_value()];
-
-        assert_round_trip_eq(&values);
-    }
-
-    #[test]
-    fn i32_round_trip() {
-        let values = [i32::min_value(), -1, 0, 1, i32::max_value()];
-
-        assert_round_trip_eq(&values);
-    }
+    // Tests for the other types.
 
     #[test]
-    fn u64_round_trip() {
-        let values = [u64::min_value(), 1, u64::max_value()];
-
-        assert_round_trip_eq(&values);
-    }
-
-    #[test]
-    fn i64_round_trip() {
-        let values = [i64::min_value(), -1, 0, 1, i64::max_value()];
-
-        assert_round_trip_eq(&values);
-    }
-
-    #[test]
-    fn bool_round_trip() {
-        let values = [false, true];
-
-        assert_round_trip_eq(&values);
-    }
-
-    #[test]
-    fn vec_round_trip() {
+    fn test_binary_form_vec_u8() {
         let values = [vec![], vec![1], vec![1, 2, 3], vec![255; 100]];
-
         assert_round_trip_eq(&values);
     }
 
     #[test]
-    fn string_round_trip() {
+    fn test_binary_form_bool_correct() {
+        let values = [true, false];
+        assert_round_trip_eq(&values);
+    }
+
+    #[test]
+    #[should_panic(expected = "Invalid value for bool: 2")]
+    fn test_binary_form_bool_incorrect() {
+        let bytes = 2_u8.to_bytes();
+        <bool as BinaryValue>::from_bytes(bytes.into()).unwrap();
+    }
+
+    #[test]
+    fn test_binary_form_string() {
         let values: Vec<_> = ["", "e", "2", "hello"]
             .iter()
             .map(|v| v.to_string())
             .collect();
-
         assert_round_trip_eq(&values);
     }
 
     #[test]
-    fn storage_value_for_system_time_round_trip() {
+    fn test_binary_form_datetime() {
         use chrono::TimeZone;
 
         let times = [
@@ -390,23 +356,21 @@ mod tests {
             Utc.timestamp(0, 999_999_999),
             Utc.timestamp(0, 1_500_000_000), // leap second
         ];
-
         assert_round_trip_eq(&times);
     }
 
     #[test]
-    fn uuid_round_trip() {
+    fn test_binary_form_uuid() {
         let values = [
             Uuid::nil(),
             Uuid::parse_str("936DA01F9ABD4d9d80C702AF85C822A8").unwrap(),
             Uuid::parse_str("0000002a-000c-0005-0c03-0938362b0809").unwrap(),
         ];
-
         assert_round_trip_eq(&values);
     }
 
     #[test]
-    fn decimal_round_trip() {
+    fn test_binary_form_decimal() {
         let values = [
             Decimal::from_str("3.14").unwrap(),
             Decimal::from_parts(1102470952, 185874565, 1703060790, false, 28),
@@ -414,17 +378,6 @@ mod tests {
             Decimal::from_str("0").unwrap(),
             Decimal::from_str("-0.000000000000000000019").unwrap(),
         ];
-
         assert_round_trip_eq(&values);
-    }
-
-    fn assert_round_trip_eq<T: StorageValue + Clone + PartialEq + Debug>(values: &[T]) {
-        for value in values.into_iter() {
-            let bytes = value.clone().into_bytes();
-            assert_eq!(
-                *value,
-                <T as StorageValue>::from_bytes(Cow::Borrowed(&bytes))
-            );
-        }
     }
 }