+++ CHANGELOG.md
@@ -10,10 +10,19 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   specification. Leaf nodes contain hashes with 0x00 prefix, branch nodes - with
   0x01.
 
+- `StorageValue` and `StorageKey` have been renamed to the `BinaryValue`
+  and `BinaryKey`. (#4)
+
+  - Added `to_bytes` method to the `BinaryValue` trait which doesn't consume
+    original value instead of the `into_bytes`.
+  - `BinaryKey::write` now returns total number of written bytes.
+  - `CryptoHash` has been replaced by the `UniqueHash`.
+
 - Changed the hash algorithm of the intermediate nodes in `ProofMapIndex`. (#1)
 
-  `ProofPath` now uses compact binary representation in the `BranchNode` hash calculation.
-
+  `ProofPath` now uses compact binary representation in the `BranchNode`
+  hash calculation.
+  
   Binary representation is `|bits_len|bytes|`, where:
 
   - **bits_len** - total length of the given `ProofPath` in bits compressed